NFT-137 update Drops Delete API
- Delete drop after checking status(only REGISTERING, NOT_SCHEDULED)
- Update drop after checking status(only REGISTERING, NOT_SCHEDULED)
- add NFT-137 e2e test code

+++ src/drops/drops.service.ts
@@ -1,8 +1,8 @@
-import { Injectable, NotFoundException, UnprocessableEntityException } from '@nestjs/common';
+import { Injectable, Logger, NotFoundException, UnprocessableEntityException } from '@nestjs/common';
 import { JwtService } from '@nestjs/jwt';
 import { InjectModel } from '@nestjs/mongoose';
 import { PaginateModel } from 'mongoose';
-import { CreateDropDto } from './dto/create-drop.dto';
+import { CreateDropDto, DropStatus } from './dto/create-drop.dto';
 import { UpdateDropDto } from './dto/update-drop.dto';
 import { Drops, DropsDocument } from './schemas/drops.schema';
 
@@ -21,14 +21,6 @@ export class DropsService {
     }
     const newCreator = new this.dropsModel(createDropDto);
     return newCreator.save();
-  }
-
-  private async checkDropExists(dropName: string): Promise<boolean> {
-    const drop = await this.dropsModel.findOne({ dropName });
-    if (drop) {
-      return true;
-    }
-    return false;
   }
 
   // Retieve creatorss with pagination
@@ -52,7 +44,11 @@ export class DropsService {
     return result;
   }
 
-  update(updateCreatorsDto: UpdateDropDto) {
+  async update(updateCreatorsDto: UpdateDropDto) {
+    if (await this.checkDropCanBeModified(updateCreatorsDto._id) === false) {
+      throw new UnprocessableEntityException('Drop can be modified only when status is registering or not scheduled');
+    }
+
     // update data into mongodb with mongoose
     return this.dropsModel.updateOne(
       { _id: updateCreatorsDto._id },
@@ -61,15 +57,38 @@ export class DropsService {
   }
 
   async remove(_id: string) {
+    if (await this.dropsModel.findById({ _id }) === null) {
+      throw new NotFoundException('drop not found');
+    }
+
+    if (await this.checkDropCanBeModified(_id) === false) {
+      throw new UnprocessableEntityException('Drop can be modified only when status is registering or not scheduled');
+    }
+
+    this.dropsModel.findById(_id);
     // delete data into mongodb with mongoose
     const result = await this.dropsModel.findByIdAndDelete(
       { _id }
     ).exec();
 
-    if (!result) {
-      throw new NotFoundException('drop not found');
-    }
-
+    Logger.debug(result);
     return `removes a ${_id} drop`;
   }
+
+  private async checkDropExists(dropName: string): Promise<boolean> {
+    const drop = await this.dropsModel.findOne({ dropName });
+    if (drop) {
+      return true;
+    }
+    return false;
+  }
+
+  private async checkDropCanBeModified(_id: string): Promise<boolean> {
+    const dropStatus = await this.dropsModel.findOne({ _id }).select({ status: 1 });
+    if (dropStatus.status === DropStatus.REGISTERING || dropStatus.status === DropStatus.NOT_SCHEDULED) {
+      return true;
+    }
+    return false;
+  }
+
 }
